Back-end/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
filelock==3.0.12
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.2.1
wrapt==1.12.1
